# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: webcup
  description: Best API REST building with GoaDesign
  contact:
    name: Guillaume MORIN
    email: guillaume.morin@epitech.eu
    url: https://guillaume-morin.fr/
  license:
    name: GOA X GM
    url: https://guillaume-morin.fr/
  version: "1.0"
host: :0
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /v1/web/public/user/{username}:
    get:
      tags:
      - publicUsers
      summary: getUserByUsername publicUsers
      description: Get one User by username
      operationId: publicUsers#getUserByUsername
      parameters:
      - name: username
        in: path
        required: true
        type: string
        maxLength: 20
        minLength: 3
        pattern: ^[a-z0-9_\-]+$
      - name: Authorization
        in: header
        description: OAuth token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/PublicUsersGetUserByUsernameResponseBody'
            required:
            - user
            - success
        "403":
          description: Forbidden response.
          schema:
            type: string
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/PublicUsersGetUserByUsernameUnknownErrorResponseBody'
            required:
            - err
            - success
            - error_code
      schemes:
      - http
      security:
      - OAuth2_header_Authorization: []
  /v1/web/public/user/recents/{offset}/{limit}:
    get:
      tags:
      - publicUsers
      summary: listUsersMostRecent publicUsers
      description: List users the most recent
      operationId: publicUsers#listUsersMostRecent
      parameters:
      - name: offset
        in: path
        description: Offset for pagination
        required: true
        type: integer
        minimum: 0
      - name: limit
        in: path
        description: Limit of items listed for pagination
        required: true
        type: integer
        minimum: 0
      - name: Authorization
        in: header
        description: OAuth token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/PublicUsersListUsersMostRecentResponseBody'
            required:
            - users
            - success
            - count
        "403":
          description: Forbidden response.
          schema:
            type: string
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/PublicUsersListUsersMostRecentUnknownErrorResponseBody'
            required:
            - err
            - success
            - error_code
      schemes:
      - http
      security:
      - OAuth2_header_Authorization: []
  /v1/web/public/user/search:
    patch:
      tags:
      - publicUsers
      summary: listUsers publicUsers
      description: List users for search bar
      operationId: publicUsers#listUsers
      parameters:
      - name: Authorization
        in: header
        description: OAuth token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      - name: ListUsersRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/PublicUsersListUsersRequestBody'
          required:
          - key
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/PublicUsersListUsersResponseBody'
            required:
            - users
            - success
        "403":
          description: Forbidden response.
          schema:
            type: string
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/PublicUsersListUsersUnknownErrorResponseBody'
            required:
            - err
            - success
            - error_code
      schemes:
      - http
      security:
      - OAuth2_header_Authorization: []
...
definitions:
...
  PublicUsersGetUserByUsernameResponseBody:
    title: PublicUsersGetUserByUsernameResponseBody
    type: object
    properties:
      success:
        type: boolean
        example: false
      user:
        $ref: '#/definitions/resUserResponseBody'
    example:
      success: false
      user:
        avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
        email: guillaume@gmail.com
        firstname: Guillaume
        id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        lastname: Morin
        role: Maiores et non dolor facilis et voluptatem.
        username: guillaumemorin
    required:
    - user
    - success
  PublicUsersGetUserByUsernameUnknownErrorResponseBody:
    title: PublicUsersGetUserByUsernameUnknownErrorResponseBody
    type: object
    properties:
      err:
        type: string
        example: sql no rows affected
      error_code:
        type: string
        example: TX_UPDATE_ITEM
      success:
        type: boolean
        default: false
        example: false
    description: Error not identified (500)
    example:
      err: sql no rows affected
      error_code: TX_UPDATE_ITEM
      success: true
    required:
    - err
    - success
    - error_code
  PublicUsersListUsersMostRecentResponseBody:
    title: PublicUsersListUsersMostRecentResponseBody
    type: object
    properties:
      count:
        type: integer
        description: total of users
        example: 6415823379680100346
        format: int64
      success:
        type: boolean
        example: true
      users:
        type: array
        items:
          $ref: '#/definitions/resUserResponseBody'
        description: Result is an an array of user
        example:
        - avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
          email: guillaume@gmail.com
          firstname: Guillaume
          id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
          lastname: Morin
          role: Maiores et non dolor facilis et voluptatem.
          username: guillaumemorin
        - avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
          email: guillaume@gmail.com
          firstname: Guillaume
          id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
          lastname: Morin
          role: Maiores et non dolor facilis et voluptatem.
          username: guillaumemorin
    example:
      count: 5321009824262484858
      success: true
      users:
      - avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
        email: guillaume@gmail.com
        firstname: Guillaume
        id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        lastname: Morin
        role: Maiores et non dolor facilis et voluptatem.
        username: guillaumemorin
      - avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
        email: guillaume@gmail.com
        firstname: Guillaume
        id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        lastname: Morin
        role: Maiores et non dolor facilis et voluptatem.
        username: guillaumemorin
      - avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
        email: guillaume@gmail.com
        firstname: Guillaume
        id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        lastname: Morin
        role: Maiores et non dolor facilis et voluptatem.
        username: guillaumemorin
      - avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
        email: guillaume@gmail.com
        firstname: Guillaume
        id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        lastname: Morin
        role: Maiores et non dolor facilis et voluptatem.
        username: guillaumemorin
    required:
    - users
    - success
    - count
  PublicUsersListUsersMostRecentUnknownErrorResponseBody:
    title: PublicUsersListUsersMostRecentUnknownErrorResponseBody
    type: object
    properties:
      err:
        type: string
        example: sql no rows affected
      error_code:
        type: string
        example: TX_UPDATE_ITEM
      success:
        type: boolean
        default: false
        example: false
    description: Error not identified (500)
    example:
      err: sql no rows affected
      error_code: TX_UPDATE_ITEM
      success: false
    required:
    - err
    - success
    - error_code
  PublicUsersListUsersRequestBody:
    title: PublicUsersListUsersRequestBody
    type: object
    properties:
      key:
        type: string
        example: guillaumemoriin
    example:
      key: guillaumemoriin
    required:
    - key
  PublicUsersListUsersResponseBody:
    title: PublicUsersListUsersResponseBody
    type: object
    properties:
      success:
        type: boolean
        example: false
      users:
        type: array
        items:
          $ref: '#/definitions/resUserResponseBody'
        description: Result is an an array of user
        example:
        - avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
          email: guillaume@gmail.com
          firstname: Guillaume
          id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
          lastname: Morin
          role: Maiores et non dolor facilis et voluptatem.
          username: guillaumemorin
        - avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
          email: guillaume@gmail.com
          firstname: Guillaume
          id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
          lastname: Morin
          role: Maiores et non dolor facilis et voluptatem.
          username: guillaumemorin
    example:
      success: true
      users:
      - avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
        email: guillaume@gmail.com
        firstname: Guillaume
        id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        lastname: Morin
        role: Maiores et non dolor facilis et voluptatem.
        username: guillaumemorin
      - avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
        email: guillaume@gmail.com
        firstname: Guillaume
        id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        lastname: Morin
        role: Maiores et non dolor facilis et voluptatem.
        username: guillaumemorin
      - avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
        email: guillaume@gmail.com
        firstname: Guillaume
        id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        lastname: Morin
        role: Maiores et non dolor facilis et voluptatem.
        username: guillaumemorin
      - avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
        email: guillaume@gmail.com
        firstname: Guillaume
        id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        lastname: Morin
        role: Maiores et non dolor facilis et voluptatem.
        username: guillaumemorin
    required:
    - users
    - success
  PublicUsersListUsersUnknownErrorResponseBody:
    title: PublicUsersListUsersUnknownErrorResponseBody
    type: object
    properties:
      err:
        type: string
        example: sql no rows affected
      error_code:
        type: string
        example: TX_UPDATE_ITEM
      success:
        type: boolean
        default: false
        example: false
    description: Error not identified (500)
    example:
      err: sql no rows affected
      error_code: TX_UPDATE_ITEM
      success: false
    required:
    - err
    - success
    - error_code
...
  resUserResponseBody:
    title: resUserResponseBody
    type: object
    properties:
      avatar:
        type: string
        example: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
      email:
        type: string
        example: guillaume@gmail.com
      firstname:
        type: string
        example: Guillaume
      id:
        type: string
        example: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
      lastname:
        type: string
        example: Morin
      role:
        type: string
        description: User is admin or not
        example: Vitae et.
      username:
        type: string
        example: guillaumemorin
    example:
      avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
      email: guillaume@gmail.com
      firstname: Guillaume
      id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
      lastname: Morin
      role: Et distinctio nihil.
      username: guillaumemorin
    required:
    - id
    - email
    - username
    - firstname
    - lastname
    - avatar
    - role
securityDefinitions:
  OAuth2_header_Authorization:
    type: oauth2
    description: Use OAuth2 to authenticate
    flow: application
    tokenUrl: /authorization
    scopes:
      api:read: Provides read access
  jwt_header_jwtToken:
    type: apiKey
    description: |-
      Secures endpoint by requiring a valid JWT token retrieved via the signin endpoint. Supports scopes "api:read" and "api:write".

      **Security Scopes**:
        * `api:read`: Read-only access
        * `api:write`: Read and write access
    name: jwtToken
    in: header